it is almost working but i am giving up
this was really fun, but i have given up here. Also if you are looking at this, be aware that app stores will almost certainly reject any app that does this in the background instead of with an active intent, unless you're the default messaging app. Also if you were actaully doing this there are people who have done this much better, like klinker41, and you should use their work. I gave up because i got bored trying to figure out why pduData was empty. it is probably a 1 line fix. Breadcrumb is: logcat: persistIfRequired: empty PDU. At this point it does send an empty mms to the end user though, just a thing that says Attachment that's 40 bytes long. so that's cool.